requirements.txt
black==22.1.0
click==8.0.4
colorama==0.4.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
python-decouple==3.6
tomli==2.0.1
typing-extensions==4.1.1
